+++ src/netspresso_trainer/dataloaders/classification/local.py
@@ -4,6 +4,7 @@ import PIL.Image as Image
 
 from ..base import BaseCustomDataset
 
+
 class ClassificationCustomDataset(BaseCustomDataset):
 
     def __init__(self, conf_data, conf_augmentation, model_name, idx_to_class,
@@ -17,10 +18,10 @@ class ClassificationCustomDataset(BaseCustomDataset):
         img = self.samples[index]['image']
         target = self.samples[index]['label'] if 'label' in self.samples[index] else None
         img = Image.open(img).convert('RGB')
-        
+
         if self.transform is not None:
-            out = self.transform(conf_augmentation=self.conf_augmentation, img_size=self.conf_augmentation.img_size)(img)
-        
+            out = self.transform(conf_augmentation=self.conf_augmentation)(img)
+
         if target is None:
             target = -1  # To be ignored at cross-entropy loss
         return out['image'], target