+++ src/netspresso_trainer/metrics/base.py
@@ -1,0 +1,11 @@
+from typing import List
+
+
+class BaseMetric:
+    metric_names: List[str] = []
+
+    def __init__(self, **kwargs):
+        pass
+
+    def calibrate(self, pred, target, **kwargs):
+        pass