+++ src/netspresso_trainer/metrics/builder.py
@@ -1,9 +1,8 @@
-from typing import Any
+from typing import Any, Dict
 
 import torch
 
-from .classification import accuracy_topk
-from .segmentation import segmentation_stats
+from .registry import TASK_METRIC
 from ..utils.record import AverageMeter
 
 MODE = ['train', 'valid', 'test']
@@ -11,72 +10,36 @@ IGNORE_INDEX_NONE_AS = -100  # following PyTorch preference
 
 
 class MetricFactory:
-    def __init__(self, conf_model, **kwargs) -> None:
-        self.metric_func_dict = dict()
-
-        ignore_index = kwargs['ignore_index'] if 'ignore_index' in kwargs else None
-        self.ignore_index = ignore_index if ignore_index is not None else IGNORE_INDEX_NONE_AS
-        self.num_classes = kwargs['num_classes'] if 'num_classes' in kwargs else None
-        
+    def __init__(self, task, conf_model, **kwargs) -> None:
+        self.task = task
         self.conf_model = conf_model
-        self._build_metric()
-        self.metric_val_dict = {
-            _mode: {
+
+        assert self.task in TASK_METRIC
+        metric_cls = TASK_METRIC[self.task]
+
+        self.metric_meter_dict: Dict[str, Dict[str, AverageMeter]] = {
+            mode: {
                 metric_key: AverageMeter(metric_key, ':6.2f')
-                for metric_key in self.metric_func_dict.keys()
+                for metric_key in metric_cls.metric_names
             }
-            for _mode in MODE
+            for mode in MODE
         }
 
-    def _build_metric(self):
-        # TODO: decide metrics by arguments
-        task = self.conf_model.task
+        self.metric_fn = metric_cls(**kwargs)
 
-        if task == 'classification':
-            self.metric_func_dict['Acc@1'] = lambda pred, target: accuracy_topk(pred, target, topk=(1, ))[0]
-            self.metric_func_dict['Acc@5'] = lambda pred, target: accuracy_topk(pred, target, topk=(5, ))[0]
-        elif task == 'segmentation':
-            assert self.num_classes is not None
-            self.metric_func_dict['iou'] = \
-                lambda pred, target: segmentation_stats(pred, target,
-                                                        ignore_index=self.ignore_index,
-                                                        num_classes=self.num_classes)['iou']
-            self.metric_func_dict['pixel_acc'] = \
-                lambda pred, target: segmentation_stats(pred, target,
-                                                        ignore_index=self.ignore_index,
-                                                        num_classes=self.num_classes)['pixel_acc']
-        elif task == 'detection':
-            # FIXME: metric for detection
-            self.metric_func_dict['iou'] = \
-                lambda pred, target: segmentation_stats(pred, target,
-                                                        ignore_index=self.ignore_index,
-                                                        num_classes=self.num_classes)['iou']
-            self.metric_func_dict['pixel_acc'] = \
-                lambda pred, target: segmentation_stats(pred, target,
-                                                        ignore_index=self.ignore_index,
-                                                        num_classes=self.num_classes)['pixel_acc']
-        else:
-            raise AssertionError
+    def __call__(self, pred: torch.Tensor, target: torch.Tensor, mode='train', *args: Any, **kwargs: Any) -> Any:
 
-    def _clear(self):
-        for avg_meter in self.loss_val_dict.values():
-            avg_meter.reset()
-
-    def clear(self):
-        self._clear()
-
-    def __call__(self, pred: torch.Tensor, target: torch.Tensor, mode='train', *args: Any, **kwds: Any) -> Any:
-        _mode = mode.lower()
-        assert _mode in MODE, f"{_mode} is not defined at our mode list ({MODE})"
-        for metric_key, metric_func in self.metric_func_dict.items():
-            metric_val = metric_func(pred, target)
-            self.metric_val_dict[_mode][metric_key].update(metric_val.item())
+        metric_result_dict = self.metric_fn.calibrate(pred, target)
+        mode = mode.lower()
+        assert mode in MODE, f"{mode} is not defined at our mode list ({MODE})"
+        for metric_key in self.metric_meter_dict[mode]:
+            assert metric_key in metric_result_dict
+            self.metric_meter_dict[mode][metric_key].update(float(metric_result_dict[metric_key]))
 
     def result(self, mode='train'):
-        _mode = mode.lower()
-        return self.metric_val_dict[_mode]
+        return self.metric_meter_dict[mode.lower()]
 
 
-def build_metrics(conf_model, **kwargs):
-    metric_handler = MetricFactory(conf_model, **kwargs)
+def build_metrics(task: str, conf_model, **kwargs) -> MetricFactory:
+    metric_handler = MetricFactory(task, conf_model, **kwargs)
     return metric_handler