+++ src/netspresso_trainer/metrics/classification/__init__.py
@@ -1,1 +1,1 @@
-from .accuracy import accuracy_topk
+from .metric import ClassificationMetric