+++ src/netspresso_trainer/pipelines/base.py
@@ -55,8 +55,8 @@ class BasePipeline(ABC):
 
         if self.single_gpu_or_rank_zero:
             self.train_logger = build_logger(self.conf, self.task, self.model_name,
-                                            step_per_epoch=self.train_step_per_epoch, class_map=class_map,
-                                            num_sample_images=NUM_SAMPLES)
+                                             step_per_epoch=self.train_step_per_epoch, class_map=class_map,
+                                             num_sample_images=NUM_SAMPLES)
 
     @final
     def _is_ready(self):
@@ -64,20 +64,20 @@ class BasePipeline(ABC):
         assert self.optimizer is not None, "`self.optimizer` is not defined!"
         """Append here if you need more assertion checks!"""
         return True
-    
+
     def _save_checkpoint(self, model: nn.Module):
         result_dir = self.train_logger.result_dir
         model_path = Path(result_dir) / f"{self.task}_{self.model_name}.ckpt"
-        
+
         save_onnx(model, model_path.with_suffix(".onnx"),
-                    sample_input=torch.randn((1, 3, self.conf.augmentation.img_size, self.conf.augmentation.img_size)))
-        
+                  sample_input=torch.randn((1, 3, self.conf.augmentation.img_size, self.conf.augmentation.img_size)))
+
         if self.is_graphmodule_training:
             torch.save(model, model_path.with_suffix(".pt"))
         else:
             torch.save(model.state_dict(), model_path.with_suffix(".pth"))
             save_graphmodule(model, (model_path.parent / f"{model_path.stem}_fx").with_suffix(".pt"))
-        
+
     @abstractmethod
     def set_train(self):
         raise NotImplementedError
@@ -89,9 +89,13 @@ class BasePipeline(ABC):
     @abstractmethod
     def valid_step(self, batch):
         raise NotImplementedError
-    
+
     @abstractmethod
     def test_step(self, batch):
+        raise NotImplementedError
+
+    @abstractmethod
+    def get_metric_with_all_outputs(self, outputs):
         raise NotImplementedError
 
     def train(self):
@@ -104,13 +108,13 @@ class BasePipeline(ABC):
         for num_epoch in range(START_EPOCH_ZERO_OR_ONE, self.conf.training.epochs + START_EPOCH_ZERO_OR_ONE):
             self.timer.start_record(name=f'train_epoch_{num_epoch}')
             self.loss = build_losses(self.conf.model, ignore_index=self.ignore_index)
-            self.metric = build_metrics(self.conf.model, ignore_index=self.ignore_index, num_classes=self.num_classes)
+            self.metric = build_metrics(self.task, self.conf.model, ignore_index=self.ignore_index, num_classes=self.num_classes)
 
             self.train_one_epoch()
 
             with_valid_logging = self.epoch_with_valid_logging(num_epoch)
             # FIXME: multi-gpu sample counting & validation
-            validation_samples = self.validate() if with_valid_logging else None
+            valid_samples = self.validate() if with_valid_logging else None
 
             self.timer.end_record(name=f'train_epoch_{num_epoch}')
             time_for_epoch = self.timer.get(name=f'train_epoch_{num_epoch}', as_pop=False)
@@ -118,7 +122,7 @@ class BasePipeline(ABC):
             if self.single_gpu_or_rank_zero:
                 self.log_end_epoch(epoch=num_epoch,
                                    time_for_epoch=time_for_epoch,
-                                   validation_samples=validation_samples,
+                                   valid_samples=valid_samples,
                                    valid_logging=with_valid_logging)
 
             self.scheduler.step()  # call after reporting the current `learning_rate`
@@ -132,7 +136,7 @@ class BasePipeline(ABC):
             self.train_logger.log_end_of_traning(final_metrics={'time_for_last_epoch': time_for_epoch})
 
             model = self.model.module if hasattr(self.model, 'module') else self.model
-            
+
             self._save_checkpoint(model)
 
     def train_one_epoch(self):
@@ -143,13 +147,17 @@ class BasePipeline(ABC):
     def validate(self, num_samples=NUM_SAMPLES):
         num_returning_samples = 0
         returning_samples = []
+        outputs = []
         for idx, batch in enumerate(tqdm(self.eval_dataloader, leave=False)):
             out = self.valid_step(batch)
-            if out is not None and num_returning_samples < num_samples:
-                returning_samples.append(out)
-                num_returning_samples += len(out['pred'])
+            if out is not None:
+                outputs.append(out)
+                if num_returning_samples < num_samples:
+                    returning_samples.append(out)
+                    num_returning_samples += len(out['pred'])
+        self.get_metric_with_all_outputs(outputs)
         return returning_samples
-    
+
     @torch.no_grad()
     def inference(self, test_dataset):
         returning_samples = []
@@ -157,8 +165,8 @@ class BasePipeline(ABC):
             out = self.test_step(batch)
             returning_samples.append(out)
         return returning_samples
-        
-    def log_end_epoch(self, epoch, time_for_epoch, validation_samples=None, valid_logging=False):
+
+    def log_end_epoch(self, epoch, time_for_epoch, valid_samples=None, valid_logging=False):
         train_losses = self.loss.result('train')
         train_metrics = self.metric.result('train')
 
@@ -172,7 +180,7 @@ class BasePipeline(ABC):
             valid_losses=valid_losses,
             valid_metrics=valid_metrics,
             train_images=None,
-            valid_images=validation_samples,
+            valid_images=valid_samples,
             learning_rate=self.learning_rate,
             elapsed_time=time_for_epoch
         )