+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## New Features:
 
-No changes to highlight.
+- Add LR simulator (powered by gradio) with `training` configuration by `@deepkyu` in `[PR 2](https://github.com/Nota-NetsPresso/netspresso-trainer/pull/2)`
+- Add augmentation simulator (powered by gradio) with `augmentation` configuration by `@deepkyu` in `[PR 2](https://github.com/Nota-NetsPresso/netspresso-trainer/pull/2)`
+- Add LR scheduler (cosine with warm restart, step_lr) by `@deepkyu` in `[PR 2](https://github.com/Nota-NetsPresso/netspresso-trainer/pull/2)`
 
 ## Bug Fixes:
 
@@ -10,7 +12,7 @@ No changes to highlight.
 
 ## Breaking Changes:
 
-No changes to highlight.
+- Support detection training with its metric by `@deepkyu` in `[PR 2](https://github.com/Nota-NetsPresso/netspresso-trainer/pull/2)`
 
 ## Other Changes: